requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pyzmq>=26.0,<27.0
psutil>=6.0,<7.0
Pillow>=10.4,<12.0
